DDD Part 4: ReviewEntity

+++ api/Concertable.Application/Requests/ReviewRequests.cs
@@ -3,6 +3,6 @@ namespace Concertable.Application.Requests;
 public record CreateReviewRequest
 {
     public int ConcertId { get; set; }
-    public int Stars { get; set; }
+    public byte Stars { get; set; }
     public string? Details { get; set; }
 }
